Refactor organise project files
